requirements.txt
moviepy
google-generativeai
pandas
edge_tts
pydub
speechrecognition
selenium
google-api-python-client